seperate compact from serialization.h
